pyproject.toml
[tool.poetry]
name = "gi-notifier"
version = "0.1.0"
description = "Notifier for Genshin Impact"
authors = ["Akashdeep Dhar <akashdeep.dhar@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/gridhead/gi-notifier"
repository = "https://github.com/gridhead/gi-notifier"
documentation = "https://github.com/gridhead/gi-notifier/blob/main/README.md"

[tool.poetry.dependencies]
python = ">=3.12"
python-telegram-bot = "^22.1"
apscheduler = "^3.11.0"
click = "^8.2.1"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.2.0 || ^0.3.0 || ^0.6.0 || ^0.7.0 || ^0.11.0"

[tool.ruff]
line-length = 100
fix = true

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"gi_notifier/data/*" = ["E501"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gi-notifier = "gi_notifier.main:main"
